+++ Sources/MapboxNavigationCore/Settings/Configuration/UnitOfMeasurement.swift
@@ -1,6 +1,7 @@
 import Foundation
 
 /// Holds available types of measurement units.
+@available(*, deprecated, message: "Use `UnitMeasurementSystem` instead.")
 public enum UnitOfMeasurement: Equatable, Sendable {
     /// Allows SDK to pick proper units.
     case auto