+++ Sources/MapboxNavigationUIKit/Utils/Locale+Conversions.swift
@@ -1,0 +1,9 @@
+import Foundation
+import MapboxDirections
+import MapboxNavigationCore
+
+extension Locale {
+    var unitMeasurementSystem: UnitMeasurementSystem {
+        return .init(locale: self)
+    }
+}